requirements.txt
Django==4.1.2
django-import_export~=3.1.0
django-pure-pagination

dj-database-url
dj-static
python-decouple
django-environ
gunicorn
psycopg2-binary
whitenoise

matplotlib